Move uri_parser to common spot
