+++ grpc.gemspec
@@ -207,6 +207,7 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/filters/http/client/http_client_filter.h )
   s.files += %w( src/core/ext/filters/http/message_compress/message_compress_filter.h )
   s.files += %w( src/core/ext/filters/http/server/http_server_filter.h )
+  s.files += %w( src/core/ext/uri/uri_parser.h )
   s.files += %w( src/core/ext/filters/client_channel/lb_policy/grpclb/grpclb.h )
   s.files += %w( src/core/ext/filters/client_channel/lb_policy/xds/xds.h )
   s.files += %w( src/core/lib/security/context/security_context.h )
@@ -291,7 +292,6 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/filters/client_channel/retry_throttle.h )
   s.files += %w( src/core/ext/filters/client_channel/subchannel.h )
   s.files += %w( src/core/ext/filters/client_channel/subchannel_index.h )
-  s.files += %w( src/core/ext/filters/client_channel/uri_parser.h )
   s.files += %w( src/core/ext/filters/deadline/deadline_filter.h )
   s.files += %w( src/core/ext/filters/client_channel/health/health.pb.h )
   s.files += %w( src/core/tsi/alts_transport_security.h )
@@ -639,6 +639,7 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/filters/http/http_filters_plugin.cc )
   s.files += %w( src/core/ext/filters/http/message_compress/message_compress_filter.cc )
   s.files += %w( src/core/ext/filters/http/server/http_server_filter.cc )
+  s.files += %w( src/core/ext/uri/uri_parser.cc )
   s.files += %w( src/core/lib/http/httpcli_security_connector.cc )
   s.files += %w( src/core/lib/security/context/security_context.cc )
   s.files += %w( src/core/lib/security/credentials/alts/alts_credentials.cc )
@@ -733,7 +734,6 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/filters/client_channel/retry_throttle.cc )
   s.files += %w( src/core/ext/filters/client_channel/subchannel.cc )
   s.files += %w( src/core/ext/filters/client_channel/subchannel_index.cc )
-  s.files += %w( src/core/ext/filters/client_channel/uri_parser.cc )
   s.files += %w( src/core/ext/filters/deadline/deadline_filter.cc )
   s.files += %w( src/core/ext/filters/client_channel/health/health.pb.c )
   s.files += %w( src/core/tsi/alts_transport_security.cc )
