+++ grpc.gyp
@@ -455,6 +455,7 @@
         'src/core/ext/filters/http/http_filters_plugin.cc',
         'src/core/ext/filters/http/message_compress/message_compress_filter.cc',
         'src/core/ext/filters/http/server/http_server_filter.cc',
+        'src/core/ext/uri/uri_parser.cc',
         'src/core/lib/http/httpcli_security_connector.cc',
         'src/core/lib/security/context/security_context.cc',
         'src/core/lib/security/credentials/alts/alts_credentials.cc',
@@ -549,7 +550,6 @@
         'src/core/ext/filters/client_channel/retry_throttle.cc',
         'src/core/ext/filters/client_channel/subchannel.cc',
         'src/core/ext/filters/client_channel/subchannel_index.cc',
-        'src/core/ext/filters/client_channel/uri_parser.cc',
         'src/core/ext/filters/deadline/deadline_filter.cc',
         'src/core/ext/filters/client_channel/health/health.pb.c',
         'src/core/tsi/alts_transport_security.cc',
@@ -809,12 +809,12 @@
         'src/core/ext/filters/client_channel/retry_throttle.cc',
         'src/core/ext/filters/client_channel/subchannel.cc',
         'src/core/ext/filters/client_channel/subchannel_index.cc',
-        'src/core/ext/filters/client_channel/uri_parser.cc',
         'src/core/ext/filters/deadline/deadline_filter.cc',
         'src/core/ext/filters/client_channel/health/health.pb.c',
         'third_party/nanopb/pb_common.c',
         'third_party/nanopb/pb_decode.c',
         'third_party/nanopb/pb_encode.c',
+        'src/core/ext/uri/uri_parser.cc',
         'src/core/ext/transport/chttp2/transport/bin_decoder.cc',
         'src/core/ext/transport/chttp2/transport/bin_encoder.cc',
         'src/core/ext/transport/chttp2/transport/chttp2_plugin.cc',
@@ -1048,12 +1048,12 @@
         'src/core/ext/filters/client_channel/retry_throttle.cc',
         'src/core/ext/filters/client_channel/subchannel.cc',
         'src/core/ext/filters/client_channel/subchannel_index.cc',
-        'src/core/ext/filters/client_channel/uri_parser.cc',
         'src/core/ext/filters/deadline/deadline_filter.cc',
         'src/core/ext/filters/client_channel/health/health.pb.c',
         'third_party/nanopb/pb_common.c',
         'third_party/nanopb/pb_decode.c',
         'third_party/nanopb/pb_encode.c',
+        'src/core/ext/uri/uri_parser.cc',
         'src/core/ext/transport/chttp2/transport/bin_decoder.cc',
         'src/core/ext/transport/chttp2/transport/bin_encoder.cc',
         'src/core/ext/transport/chttp2/transport/chttp2_plugin.cc',
@@ -1272,6 +1272,7 @@
         'src/core/ext/filters/http/http_filters_plugin.cc',
         'src/core/ext/filters/http/message_compress/message_compress_filter.cc',
         'src/core/ext/filters/http/server/http_server_filter.cc',
+        'src/core/ext/uri/uri_parser.cc',
         'src/core/ext/transport/chttp2/server/chttp2_server.cc',
         'src/core/ext/transport/chttp2/client/insecure/channel_create.cc',
         'src/core/ext/transport/chttp2/client/insecure/channel_create_posix.cc',
@@ -1299,7 +1300,6 @@
         'src/core/ext/filters/client_channel/retry_throttle.cc',
         'src/core/ext/filters/client_channel/subchannel.cc',
         'src/core/ext/filters/client_channel/subchannel_index.cc',
-        'src/core/ext/filters/client_channel/uri_parser.cc',
         'src/core/ext/filters/deadline/deadline_filter.cc',
         'src/core/ext/filters/client_channel/health/health.pb.c',
         'third_party/nanopb/pb_common.c',
