+++ src/core/ext/filters/client_channel/parse_address.h
@@ -23,7 +23,7 @@
 
 #include <stddef.h>
 
-#include "src/core/ext/filters/client_channel/uri_parser.h"
+#include "src/core/ext/uri/uri_parser.h"
 #include "src/core/lib/iomgr/resolve_address.h"
 
 /** Populate \a resolved_addr from \a uri, whose path is expected to contain a
